+++ bluefin/CHANGELOG.md
@@ -1,3 +1,5 @@
+* Add `hGetLine` and `hIsEOF` to `Bluefin.System.IO`
+
 ## 0.0.12.0
 
 * Add `asks` and `local` to `Bluefin.Reader`